remove function remove from listview

+++ SqliteSample/app/page/main/main-view-model.ts
@@ -9,14 +9,15 @@ import * as lodash from 'lodash';
 export class MainViewModel extends Observable {
     Person: PersonModel;
     PersonList: ObservableArray<PersonModel> = new ObservableArray<PersonModel>();
+
     constructor() {
         super();
         this.Person = new PersonModel();
     }
 
-    onItemTap(args: ItemEventData){
+    onItemTap(args: ItemEventData) {
         this.Person = this.PersonList.getItem(args.index);
-         this.set("Person", this.Person);
+        this.notify({ object: this, eventName: Observable.propertyChangeEvent, propertyName: 'Person', value: this.Person });
     }
 
     insert() {
@@ -28,7 +29,7 @@ export class MainViewModel extends Observable {
         });
     }
 
-    select() {
+    select(isUpdated: boolean = false) {
         Database.select(this.Person.TableName).then(rows => {
             for (var i = 0; i < rows.length; i++) {
                 let model = new PersonModel();
@@ -38,10 +39,17 @@ export class MainViewModel extends Observable {
                 model.CreatedAt = rows[i]["CreatedAt"];
                 if ((this.PersonList.filter(person => person.ID == model.ID)).length === 0) {
                     this.PersonList.push(model);
-                }
-                console.log("rows ", rows.length);
-            }
+                    console.log("pato");
 
+                } else if (isUpdated == true) {
+                    this.PersonList.setItem(this.PersonList.indexOf(this.Person), model);
+                    console.log("pato2");
+
+                }
+            }
+            this.Person.firstname = '';
+            this.Person.lastname = '';
+            this.notify({ object: this, eventName: Observable.propertyChangeEvent, propertyName: 'Person', value: this.Person });
             this.set("PersonList", this.PersonList);
         }, error => {
             console.log("select error", error);
@@ -49,18 +57,33 @@ export class MainViewModel extends Observable {
     }
 
     remove() {
-        Database.remove(this.Person.TableName, `firstname = '${this.Person.firstname}'`).then(id => {
-            console.log("remove success", id);
-        }, error => {
-            console.log("remove error", error);
-        });
+        if (this.Person.firstname != '') {
+            Database.remove(this.Person.TableName, `firstname = '${this.Person.firstname}'`).then(id => {
+                console.log("remove success", id);
+                for (var i = 0; i < this.PersonList.length; i++) {
+                    if (this.PersonList.getItem(i) === this.Person) {
+                        this.PersonList.splice(i, 1);
+                        break;
+                    }
+                }
+                this.Person.firstname = '';
+                this.Person.lastname = '';
+                this.notify({ object: this, eventName: Observable.propertyChangeEvent, propertyName: 'Person', value: this.Person });
+                this.set("PersonList", this.PersonList);
+            }, error => {
+                console.log("remove error", error);
+            });
+        }
     }
 
     update() {
-        Database.update(this.Person, `where firstname = '${this.Person.firstname}'`).then(id => {
-            console.log("update success", id);
-        }, error => {
-            console.log("update error", error);
-        });
+        if (this.Person.firstname) {
+            Database.update(this.Person, `where firstname = '${this.Person.firstname}'`).then(id => {
+                console.log("update success", id);
+                this.select(true);
+            }, error => {
+                console.log("update error", error);
+            });
+        }
     }
 }
